admittance_controller: resolve implicit cast errors

+++ admittance_controller/include/admittance_controller/admittance_rule.hpp
@@ -63,10 +63,10 @@ struct AdmittanceState
     mass_inv.setZero();
     stiffness.setZero();
     selected_axes.setZero();
-    current_joint_pos = Eigen::VectorXd::Zero(num_joints);
-    joint_pos = Eigen::VectorXd::Zero(num_joints);
-    joint_vel = Eigen::VectorXd::Zero(num_joints);
-    joint_acc = Eigen::VectorXd::Zero(num_joints);
+    current_joint_pos = Eigen::VectorXd::Zero(static_cast<long>(num_joints));
+    joint_pos = Eigen::VectorXd::Zero(static_cast<long>(num_joints));
+    joint_vel = Eigen::VectorXd::Zero(static_cast<long>(num_joints));
+    joint_acc = Eigen::VectorXd::Zero(static_cast<long>(num_joints));
   }
 
   Eigen::VectorXd current_joint_pos;
